refactor: Update import order and remove unused styles in UseMovieFetch.ts and useHomeFetch.ts
