+++ src/components/Actor/actor.props.ts
@@ -1,0 +1,5 @@
+export type Props = {
+  name: string;
+  character: string;
+  imageUrl: string;
+};
