+++ src/hooks/props.ts
@@ -1,23 +1,6 @@
-export type MoviePropTypes = {
-  adult: boolean;
-  backdrop_path: string;
-  budget: number;
-  genres: { id: number; name: string }[];
-  homepage: string;
-  id: number;
-  logo_path: string;
-  original_language: string;
-  original_title: string;
-  overview: string;
-  popularity: number;
-  poster_path: string;
-  production_companies: { name: string; id: number; logo_path: string }[];
-  release_date: string;
-  revenue: number;
-  runtime: number;
-  status: string;
-  tagline: string;
-  title: string;
-  vote_average: number;
-  vote_count: number;
+export type Cast = {
+  character: string;
+  credit_id: string;
+  name: string;
+  profile_path: string;
 };
